# :sparkles: Faster uuid generation