--- a/aries_cloudagent/indy/credx/holder.py
+++ b/aries_cloudagent/indy/credx/holder.py
@@ -4,7 +4,6 @@ import asyncio
 import json
 import logging
 import re
-import uuid
 from typing import Dict, Optional, Sequence, Tuple, Union
 
 from aries_askar import AskarError, AskarErrorCode
@@ -17,6 +16,7 @@ from indy_credx import (
     Presentation,
     PresentCredentials,
 )
+from uuid_utils import uuid4
 
 from ...askar.profile import AskarProfile
 from ...ledger.base import BaseLedger
@@ -195,7 +195,7 @@ class IndyCredxHolder(IndyHolder):
                 f"Error parsing credential definition ID: {cred_def_id}"
             )
 
-        credential_id = credential_id or str(uuid.uuid4())
+        credential_id = credential_id or str(uuid4())
         tags = {
             "schema_id": schema_id,
             "schema_issuer_did": schema_id_parts[1],