--- a/aries_cloudagent/ledger/multiple_ledger/ledger_config_schema.py
+++ b/aries_cloudagent/ledger/multiple_ledger/ledger_config_schema.py
@@ -1,8 +1,7 @@
 """Schema for configuring multiple ledgers."""
 
-import uuid
-
 from marshmallow import EXCLUDE, fields, pre_load
+from uuid_utils import uuid4
 
 from ...messaging.models.base import BaseModel, BaseModelSchema
 from ...messaging.models.openapi import OpenAPISchema
@@ -56,7 +55,7 @@ class LedgerConfigInstanceSchema(BaseModelSchema):
     def validate_id(self, data, **kwargs):
         """Check if id is present, if not then set to UUID4."""
         if "id" not in data:
-            data["id"] = str(uuid.uuid4())
+            data["id"] = str(uuid4())
         return data