--- a/aries_cloudagent/protocols/out_of_band/v1_0/manager.py
+++ b/aries_cloudagent/protocols/out_of_band/v1_0/manager.py
@@ -4,7 +4,8 @@ import asyncio
 import logging
 import re
 from typing import List, Mapping, NamedTuple, Optional, Sequence, Text, Union
-import uuid
+
+from uuid_utils import uuid4
 
 from aries_cloudagent.protocols.coordinate_mediation.v1_0.route_manager import (
     RouteManager,
@@ -145,7 +146,7 @@ class InvitationCreator:
         self.route_manager = route_manager
         self.oob = oob
 
-        self.msg_id = str(uuid.uuid4())
+        self.msg_id = str(uuid4())
         self.attachments = attachments
 
         self.handshake_protocols = [