--- a/aries_cloudagent/revocation/indy.py
+++ b/aries_cloudagent/revocation/indy.py
@@ -18,7 +18,6 @@ from ..protocols.endorse_transaction.v1_0.util import (
     is_author_role,
 )
 from ..storage.base import StorageNotFoundError
-
 from .error import (
     RevocationError,
     RevocationInvalidStateValueError,