--- a/aries_cloudagent/revocation/models/issuer_rev_reg_record.py
+++ b/aries_cloudagent/revocation/models/issuer_rev_reg_record.py
@@ -3,7 +3,6 @@
 import json
 import importlib
 import logging
-import uuid
 from functools import total_ordering
 from os.path import join
 from pathlib import Path
@@ -12,6 +11,7 @@ from typing import Any, Mapping, Sequence, Tuple, Union
 from urllib.parse import urlparse
 
 from marshmallow import fields, validate
+from uuid_utils import uuid4
 
 from ...core.profile import Profile, ProfileSession
 from ...indy.credx.issuer import (
@@ -190,7 +190,7 @@ class IssuerRevRegRecord(BaseRecord):
     async def generate_registry(self, profile: Profile):
         """Create the revocation registry definition and tails file."""
         if not self.tag:
-            self.tag = self._id or str(uuid.uuid4())
+            self.tag = self._id or str(uuid4())
 
         if self.state != IssuerRevRegRecord.STATE_INIT:
             raise RevocationError(