--- a/aries_cloudagent/transport/inbound/manager.py
+++ b/aries_cloudagent/transport/inbound/manager.py
@@ -1,9 +1,10 @@
 """Inbound transport manager."""
 
 import logging
-import uuid
 from collections import OrderedDict
 from typing import Callable, Coroutine
+
+from uuid_utils import uuid4
 
 from ...core.profile import Profile
 from ...utils.classloader import ClassLoader, ModuleLoadError, ClassNotFoundError
@@ -164,7 +165,7 @@ class InboundTransportManager:
             client_info=client_info,
             close_handler=self.closed_session,
             inbound_handler=self.receive_inbound,
-            session_id=str(uuid.uuid4()),
+            session_id=str(uuid4()),
             transport_type=transport_type,
             wire_format=wire_format,
         )