--- a/aries_cloudagent/transport/inbound/manager.py
+++ b/aries_cloudagent/transport/inbound/manager.py
@@ -7,12 +7,10 @@ from typing import Callable, Coroutine
 from uuid_utils import uuid4
 
 from ...core.profile import Profile
-from ...utils.classloader import ClassLoader, ModuleLoadError, ClassNotFoundError
+from ...utils.classloader import ClassLoader, ClassNotFoundError, ModuleLoadError
 from ...utils.task_queue import CompletedTask, TaskQueue
-
 from ..outbound.message import OutboundMessage
 from ..wire_format import BaseWireFormat
-
 from .base import (
     BaseInboundTransport,
     InboundTransportConfiguration,