@ff137 2024-05-30
not sure of the mean/variance for the integration tests, but it seems to consistently shave some time off